+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Add support for custom installations
+
 ## [0.2.0] - 2022-06-20
 
 - Add `createDatasource`